+++ docs/source/changelog.rst
@@ -1,6 +1,28 @@
 Changelog
 =========
 
+Version 0.6.0
+=============
+
+* **API-breaking**: Changed requester-pays handling for ``GCSFileSystem``.
+
+  The ``user_project`` keyword has been removed, and has been replaced with
+  the ``requester_pays`` keyword. If you're working with a ``requester_pays`` bucket
+  you will need to explicity pass ``requester_pays=True``. This will include your
+  ``project`` ID in requests made to GCS.
+
+Version 0.5.3
+=============
+
+* ``GCSFileSystem`` now validates that the ``project`` provided, if any, matches the
+  Google default project when using ``token='google_default'`` to authenticate (:pr:`219`).
+* Fixed bug in ``GCSFileSystem.cat`` on objects in requester-pays buckets (:pr:`217`).
+
+Version 0.5.2
+=============
+
+* Fixed bug in ``user_project`` fallback for default Google authentication (:pr:`213`)
+
 Version 0.5.1
 =============